add command to import data
