+++ requirements.txt
@@ -1,3 +1,4 @@
 bs4==0.0.1
 Django==1.9.7
 djangorestframework==3.3.3
+requests==2.10.0
